Create ValidPalindrome.java
A palindrome is a phrase that reads the same forward and backward after converting all uppercase letters to lowercase and removing non-alphanumeric characters. Non-alphanumeric characters include anything other than letters and numbers.

Determine if the given string "s" is a palindrome and return true if it is, or false if it's not.