ValidAnagram.java
Determine whether two strings, "s" and "t," are anagrams of each other. Return true if "t" is an anagram of "s," and false otherwise.

An anagram is a word or phrase that is formed by rearranging the letters of a different word or phrase, while using all the original letters exactly once.